requirements.txt
asgiref==3.4.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.8
Django==3.2.9
django-crispy-forms==1.13.0
django-model-utils==4.2.0
django-money==2.1
fontawesome-free==5.15.4
idna==3.3
iniconfig==1.1.1
mysqlclient==2.1.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
py-moneyed==1.2
pyparsing==3.0.6
pytest==6.2.5
pytest-django==4.5.1
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
